+++ dwds/lib/src/loaders/require.dart
@@ -123,8 +123,7 @@ class RequireStrategy extends LoadStrategy {
     this._sourceMapPathForModule,
     this._serverPathForAppUri,
     AssetReader assetReader,
-    LogWriter logWriter,
-  ) : super(assetReader, logWriter);
+  ) : super(assetReader);
 
   @override
   Handler get handler => (request) async {