+++ dwds/test/fixtures/context.dart
@@ -9,7 +9,6 @@ import 'dart:io';
 import 'package:build_daemon/client.dart';
 import 'package:build_daemon/data/build_status.dart';
 import 'package:build_daemon/data/build_target.dart';
-import 'package:build_daemon/data/server_log.dart' as server_log;
 import 'package:dwds/dwds.dart';
 import 'package:dwds/src/debugging/webkit_debugger.dart';
 import 'package:dwds/src/loaders/frontend_server_require.dart';
@@ -19,7 +18,6 @@ import 'package:dwds/src/utilities/dart_uri.dart';
 import 'package:dwds/src/utilities/shared.dart';
 import 'package:frontend_server_common/src/resident_runner.dart';
 import 'package:http/http.dart';
-import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
 import 'package:shelf/shelf.dart';
 import 'package:shelf_proxy/shelf_proxy.dart';
@@ -105,7 +103,6 @@ class TestContext {
       bool restoreBreakpoints,
       CompilationMode compilationMode,
       bool enableExpressionEvaluation,
-      LogWriter logWriter,
       bool verbose}) async {
     reloadConfiguration ??= ReloadConfiguration.none;
     serveDevTools ??= false;
@@ -115,7 +112,6 @@ class TestContext {
     waitToDebug ??= false;
     compilationMode ??= CompilationMode.buildDaemon;
     enableExpressionEvaluation ??= false;
-    logWriter ??= (Level level, String message) => printOnFailure(message);
     spawnDds ??= true;
     verbose ??= false;
 
@@ -161,11 +157,8 @@ class TestContext {
               ],
               if (verbose) '--verbose',
             ];
-            daemonClient = await connectClient(
-                workingDirectory,
-                options,
-                (log) => logWriter(
-                    server_log.toLoggingLevel(log.level), log.message));
+            daemonClient = await connectClient(workingDirectory, options,
+                (log) => printOnFailure(log.message));
             daemonClient.registerBuildTarget(
                 DefaultBuildTarget((b) => b..target = pathToServe));
             daemonClient.startBuild();
@@ -179,8 +172,8 @@ class TestContext {
             assetHandler = proxyHandler(
                 'http://localhost:$assetServerPort/$pathToServe/',
                 client: client);
-            assetReader = ProxyServerAssetReader(assetServerPort, logWriter,
-                root: pathToServe);
+            assetReader =
+                ProxyServerAssetReader(assetServerPort, root: pathToServe);
 
             if (enableExpressionEvaluation) {
               var ddcAssetHandler = proxyHandler(
@@ -192,15 +185,16 @@ class TestContext {
                 port,
                 pathToServe,
                 ddcAssetHandler,
-                logWriter,
                 verbose,
               );
               expressionCompiler = ddcService;
             }
 
             requireStrategy = BuildRunnerRequireStrategyProvider(
-                    assetHandler, reloadConfiguration, assetReader, logWriter)
-                .strategy;
+              assetHandler,
+              reloadConfiguration,
+              assetReader,
+            ).strategy;
 
             buildResults = daemonClient.buildResults;
           }
@@ -218,7 +212,6 @@ class TestContext {
                 [fileSystemRoot],
                 'org-dartlang-app',
                 _outputDir.path,
-                logWriter,
                 verbose);
 
             var assetServerPort = await findUnusedPort();
@@ -232,8 +225,9 @@ class TestContext {
             assetHandler = webRunner.devFS.assetServer.handleRequest;
 
             requireStrategy = FrontendServerRequireStrategyProvider(
-                    reloadConfiguration, assetReader, logWriter)
-                .strategy;
+              reloadConfiguration,
+              assetReader,
+            ).strategy;
 
             buildResults = const Stream<BuildResults>.empty();
           }
@@ -268,25 +262,25 @@ class TestContext {
       var connection = ChromeConnection('localhost', debugPort);
 
       testServer = await TestServer.start(
-          hostname,
-          port,
-          assetHandler,
-          assetReader,
-          requireStrategy,
-          pathToServe,
-          buildResults,
-          () async => connection,
-          serveDevTools,
-          enableDebugExtension,
-          autoRun,
-          enableDebugging,
-          useSse,
-          urlEncoder,
-          restoreBreakpoints,
-          expressionCompiler,
-          spawnDds,
-          ddcService,
-          logWriter);
+        hostname,
+        port,
+        assetHandler,
+        assetReader,
+        requireStrategy,
+        pathToServe,
+        buildResults,
+        () async => connection,
+        serveDevTools,
+        enableDebugExtension,
+        autoRun,
+        enableDebugging,
+        useSse,
+        urlEncoder,
+        restoreBreakpoints,
+        expressionCompiler,
+        spawnDds,
+        ddcService,
+      );
 
       appUrl = 'http://localhost:$port/$path';
       await webDriver.get(appUrl);