+++ dwds/test/frontend_server_evaluate_test.dart
@@ -6,10 +6,9 @@
 @TestOn('vm')
 import 'dart:async';
 
-import 'package:path/path.dart' as p;
-
 import 'package:dwds/src/connections/debug_connection.dart';
 import 'package:dwds/src/services/chrome_proxy_service.dart';
+import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 import 'package:vm_service/vm_service.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
@@ -32,7 +31,6 @@ void main() async {
       await context.setUp(
           enableExpressionEvaluation: true,
           compilationMode: CompilationMode.frontendServer,
-          logWriter: (level, message) => printOnFailure(message),
           verbose: false);
     });
 
@@ -198,7 +196,6 @@ void main() async {
       await context.setUp(
           enableExpressionEvaluation: false,
           compilationMode: CompilationMode.frontendServer,
-          logWriter: (level, message) => printOnFailure(message),
           verbose: false);
     });